+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # develop (2.0.0.head)
   * Feature: Add `ahn generate` command to allow invocation of generators
   * Feature: Add simple generator for call controllers
+  * Feature: Add simple generator for plugins
   * Feature: Allow plugins to register their generator classes
   * Feature: Add log level helper methods to Console
   * Feature: Console's shutdown/exit method initiates the shutdown routine
@@ -10,6 +11,10 @@
   * Bugfix: CallController#dial now blocks until all outbound calls complete
   * Bugfix: Call commands timing out now raise a timeout exception in the caller, but do not crash the actor
   * Feature: It is now possible to pause/resume call controllers
+  * Bugfix: CallController#dial now unblocks immediately if the original call ends
+  * Bugfix: CallController#dial now unblocks when the connected outbound call unjoins, rather than ending, incase post-processing on the outbound call is required
+  * Bugfix: CallController#dial now hangs up outbound legs when it unblocks
+  * Feature: CallController#dial now defaults the outbound caller ID to that of the controller's call
 
 # 2.0.0.alpha2 - 2012-01-30
   * Change: Plugins no longer load dialplan/event/rpc/console methods using corresponding class methods